Add files via upload
This a Guessing Game. 
It lets the user guess a number between 1-10 and guides him to reach the number guessed by the program by directing them to go either higher or lower, finally matching the two numbers. It displays the no of times in which they could guess it correctly and maintains the highest score in a separate text file.